+++ distributor-script/distributor-js/src/main/java/fr/xpdustry/distributor/exception/BlockingScriptError.java
@@ -1,11 +1,15 @@
 package fr.xpdustry.distributor.exception;
 
+import java.io.Serial;
 import org.jetbrains.annotations.NotNull;
 
 /**
  * This error is thrown when a script is running for too long.
  */
 public final class BlockingScriptError extends Error {
+
+  @Serial
+  private static final long serialVersionUID = 7267751075275052115L;
 
   private final int maxRuntime;