+++ distributor-script/distributor-js/src/main/resources/init.js
@@ -3,20 +3,20 @@
 "use strict";
 
 // these are not strictly necessary, but are kept for edge cases
-const run = method => new Runnable(){run: method}
-const boolf = method => new Boolf(){get: method}
-const boolp = method => new Boolp(){get: method}
-const floatf = method => new Floatf(){get: method}
-const floatp = method => new Floatp(){get: method}
-const cons = method => new Cons(){get: method}
-const prov = method => new Prov(){get: method}
-const func = method => new Func(){get: method}
+const run = method => new java.lang.Runnable(){run: method}
+const boolf = method => new Packages.arc.func.Boolf(){get: method}
+const boolp = method => new Packages.arc.func.Boolp(){get: method}
+const floatf = method => new Packages.arc.func.Floatf(){get: method}
+const floatp = method => new Packages.arc.func.Floatp(){get: method}
+const cons = method => new Packages.arc.func.Cons(){get: method}
+const prov = method => new Packages.arc.func.Prov(){get: method}
+const func = method => new Packages.arc.func.Func(){get: method}
 const runner = method => new Packages.arc.util.CommandHandler.CommandRunner(){accept: method}
 
-const debug = function(/*text, args...*/){Log.debug(arguments[0], Array.from(arguments).splice(1, arguments.length))}
-const info = function(/*text, args...*/){Log.info(arguments[0], Array.from(arguments).splice(1, arguments.length))}
-const warn = function(/*text, args...*/){Log.warn(arguments[0], Array.from(arguments).splice(1, arguments.length))}
-const err = function(/*text, args...*/){Log.err(arguments[0], Array.from(arguments).splice(1, arguments.length))}
+const debug = function(/*text, args...*/){Packages.arc.util.Log.debug(arguments[0], Array.from(arguments).splice(1, arguments.length))}
+const info = function(/*text, args...*/){Packages.arc.util.Log.info(arguments[0], Array.from(arguments).splice(1, arguments.length))}
+const warn = function(/*text, args...*/){Packages.arc.util.Log.warn(arguments[0], Array.from(arguments).splice(1, arguments.length))}
+const err = function(/*text, args...*/){Packages.arc.util.Log.err(arguments[0], Array.from(arguments).splice(1, arguments.length))}
 
 const print = text => Packages.java.lang.System.out.print(text)
 const println = text => Packages.java.lang.System.out.println(text)